Move battle move menu to top-level package
