+++ menus/__init__.py
@@ -1,0 +1,6 @@
+"""Menu package exposing common EvMenu flows."""
+
+from . import battle_move
+
+__all__ = ["battle_move"]
+
