+++ tests/test_battle_move_menu.py
@@ -4,7 +4,7 @@ import sys
 # Stub evennia evmenu so importing battle_move works without full evennia setup
 sys.modules['evennia.utils.evmenu'] = types.SimpleNamespace(EvMenuGotoAbortMessage=Exception)
 
-from utils.menus import battle_move
+from menus import battle_move
 
 
 def test_route_move_cancel_returns_cancel_node():
